Fix integer division

+++ initPherGrid.py
@@ -7,7 +7,7 @@ def initPherGrid(n, MAXPHER):
     site[0, :] = 0 #- Border
     site[-1,:] = 0 #- Border
     
-    middle_index = n / 2
+    middle_index = int(n / 2)
     for i in range(1,n+1):
         site[middle_index, i] = MAXPHER * i /n
